fix: выпилены ненужные хедеры
