--- src/db/components_db_data.h
@@ -1,12 +1,0 @@
-#pragma once
-#ifndef COMPONENTS_DB_DATA_H
-#define COMPONENTS_DB_DATA_H
-
-namespace vle_solvers
-{
-;
-    
-}
-
-
-#endif // COMPONENTS_DB_DATA_H
