+++ src/vle_solvers.h
@@ -5,8 +5,6 @@
 
 #include "db/common_db.h"
 #include "db/components_db.h"
-#include "db/components_db_data.h"
-#include "db/components_db_init.h"
 #include "db/tdb_serialize_2024_05_24.h"
 #include "db/tdb_serialize_common.h"
