+++ vle_solvers.vcxproj
@@ -141,8 +141,6 @@
   <ItemGroup>
     <ClInclude Include="src\db\common_db.h" />
     <ClInclude Include="src\db\components_db.h" />
-    <ClInclude Include="src\db\components_db_data.h" />
-    <ClInclude Include="src\db\components_db_init.h" />
     <ClInclude Include="src\db\tdb_serialize_2024_05_24.h" />
     <ClInclude Include="src\db\tdb_serialize_common.h" />
     <ClInclude Include="src\vle_solvers.h" />
@@ -150,6 +148,7 @@
   <ItemGroup>
     <ClCompile Include="src\db\components_db.cpp" />
     <ClCompile Include="src\db\components_db_data.cpp" />
+    <ClCompile Include="src\db\components_db_init.cpp" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
