+++ vle_solvers.vcxproj.filters
@@ -30,18 +30,15 @@
     <ClInclude Include="src\vle_solvers.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="src\db\components_db_data.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="src\db\components_db_init.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="src\db\components_db.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="src\db\components_db_data.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="src\db\components_db_init.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
